refactor(dsl): remove $iCycle, fully replaced by $p.s
$iCycle (the IntervalSeq scale-degree sequencer module) is superseded by
$cycle($p.s(source, scale)). Remove the module and its DSL helper; extract the
shared scale-degree value machinery $p.s depends on into interval_value.rs.

- Delete crates/modular_core/src/dsp/seq/interval_seq.rs (IntervalSeq module,
  params, state, Sampleable impl) and its constructor/deserializer/schema
  registration in seq/mod.rs.
- Extract IntervalValue (+ FromMiniAtom/HasRest impls), add/sub_interval_values
  and EMPTY_PATTERN_SOURCE_ERR into dsp/seq/interval_value.rs — these back the
  $p.s path (seq_value::from_sp_payload, sp_combine). Update its error
  messages to reference $p.s instead of $iCycle.
- Regenerate schemas.json: with the IntervalSeq schema gone, the schema-driven
  $iCycle factory and DSL global no longer exist.
- Tests: drop the $iCycle module tests and the IntervalSeq CV-hold integration
  test; convert the IntervalValue atom-rejection test to $p.s (the live
  consumer of that validation); add a $cycle($p.s(...)) executor test.
- Docs: update every comment referencing $iCycle/IntervalSeq across the seq,
  pattern_system/mini, quantizer, executor, argumentSpanAnalyzer,
  typescriptLibGen, ast and schemaTypeResolver sources.

Kept: the Help -> Migrate tooling that rewrites $iCycle(...) -> $cycle($p.s(...))
and $cycle("...") -> $cycle($p("...")). It operates on source text, so it
remains the migration path for existing buffers that still call $iCycle.

Verification: tsc clean; vitest 368/368; cargo test -p modular_core 598/598
(incl sp_combine parity, peggy parity, dsp_fresh). No new dead-code warnings
versus the parent branch.
