+++ crates/modular_core/src/pattern_system/mini/ast.rs
@@ -4,7 +4,7 @@
 //! including source span information for editor highlighting.
 //!
 //! Parsing itself lives TypeScript-side (`src/main/dsl/miniNotation/`). The
-//! DSL ships each `$cycle` / `$iCycle` pattern as a JSON `{ ast, source,
+//! DSL ships each `$cycle` / `$p.s` pattern as a JSON `{ ast, source,
 //! all_spans }` payload built by `$p(...)`; the Rust side deserializes it
 //! here and forwards the `MiniAST` to `convert::convert` to build the
 //! `Pattern<T>`.
@@ -276,8 +276,8 @@ pub enum MiniAST {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize, JsonSchema)]
 pub enum AtomValue {
     /// Numeric value. Module-level semantics decide the interpretation:
-    /// `$cycle` maps a bare `Number` to volts (1V/oct); `$iCycle` maps it
-    /// to an integer scale degree.
+    /// `$cycle($p(...))` maps a bare `Number` to volts (1V/oct); `$p.s(...)`
+    /// maps it to an integer scale degree.
     Number(f64),
 
     /// Frequency in Hz.
